Iteration (İterasyon)
Explanation (ENG)

Firstly, you must have an area (square, triangle etc.) to iteration calculation . Also, side values must be known (All values are different for each side). Thanks to this code, we can avoid doing the mathematical operations within the code every time. We do not know how many times the iteration will be done, but after every time we run the code, we can narrow the range and reach the result (total iteration). The code acts as a result of the number of transactions that must be repeated for the points inside and on the edges of the field. The purpose of this is to reach the information that a selected point will repeat the same number after the operation. For instance Iteration(first); V22. point: 15,22 , Iteration (last): V22. point: 15,22 again etc. iteration repetition ends when the last two digits after the comma are the same.

Açıklama (TR)

Öncelikle, iterasyon hesaplanmasında bir alana (kare, üçgen vb.) sahip olmalısınız. Ayrıca, kenar değerler de bilinmelidir (tüm değerler her kenar için farklıdır, aynı olması da sorun yaratmaz). Bu kod sayesinde, her zaman kodun içindeki matematiksel işlemleri yapmaktan kaçınabiliriz. Yinelemenin kaç kez gerçekleştirileceğini bilmiyoruz, ancak kodu her çalıştırmamızdan sonra aralığı daraltabilir ve sonuca ulaşabiliriz (toplam iterasyon sayısı). Kod, alanın içindeki ve kenarlarındaki noktalar için tekrarlanması gereken işlemlerin bir sonucu olarak çalışır. Bunun amacı, seçilen bir noktanın işlemden sonra aynı sonucu tekrarlayacağı bilgisine ulaşmaktır. Örneğin İterasyon (ilk); V22. nokta: 15,22, İterasyon (son): V22. nokta:15,22 vs. iterasyon tekrarı virgülden sonraki son iki rakamın aynı olmasıyla sonlanır.